requirements.txt
# PROMETHEUS Enterprise Trading Platform - Python Dependencies
# Version: 2.1 Ultimate with Global Intelligence
# Last Updated: October 12, 2025

# ============================================================================
# CORE WEB FRAMEWORK
# ============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ============================================================================
# DATABASE & ORM
# ============================================================================
sqlalchemy==2.0.23
alembic==1.13.0
aiosqlite==0.19.0
databases==0.8.0

# ============================================================================
# ASYNC & NETWORKING
# ============================================================================
aiohttp==3.9.1
aiofiles==23.2.1
websockets==12.0
httpx==0.25.2

# ============================================================================
# DATA PROCESSING & ANALYSIS
# ============================================================================
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2

# ============================================================================
# BROKER INTEGRATIONS
# ============================================================================
alpaca-py==0.13.0
alpaca-trade-api==3.1.1
ibapi==9.81.1.post1
ib-insync==0.9.86

# ============================================================================
# MARKET DATA PROVIDERS
# ============================================================================
yfinance==0.2.32
polygon-api-client==1.12.4
alpha-vantage==2.3.1

# ============================================================================
# DATA INTELLIGENCE SOURCES (NEW!)
# ============================================================================
# Google Trends
pytrends==4.9.2

# Reddit API
praw==7.7.1

# Twitter/X API
tweepy==4.14.0
textblob==0.17.1

# News Scraping
feedparser==6.0.10
beautifulsoup4==4.12.2
newspaper3k==0.2.8

# Weather Data
pyowm==3.3.0

# ============================================================================
# AI & MACHINE LEARNING
# ============================================================================
openai==1.3.7
anthropic==0.7.7
transformers==4.35.2
torch==2.1.1
tensorflow==2.15.0
keras==2.15.0

# ============================================================================
# TECHNICAL ANALYSIS
# ============================================================================
ta-lib==0.4.28
pandas-ta==0.3.14b0
mplfinance==0.12.10b0
stockstats==0.6.2

# ============================================================================
# SYSTEM MONITORING
# ============================================================================
psutil==5.9.6
py-cpuinfo==9.0.0

# ============================================================================
# UTILITIES
# ============================================================================
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3.post1
tzlocal==5.2

# ============================================================================
# LOGGING & DEBUGGING
# ============================================================================
loguru==0.7.2
colorama==0.4.6

# ============================================================================
# SECURITY & AUTHENTICATION
# ============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# ============================================================================
# TESTING (Optional)
# ============================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# ============================================================================
# DATA COMPRESSION & SERIALIZATION
# ============================================================================
lz4==4.3.2
msgpack==1.0.7
orjson==3.9.10

# ============================================================================
# SCHEDULING & BACKGROUND TASKS
# ============================================================================
apscheduler==3.10.4
celery==5.3.4

# ============================================================================
# CONFIGURATION & VALIDATION
# ============================================================================
pyyaml==6.0.1
toml==0.10.2
jsonschema==4.20.0

# ============================================================================
# PERFORMANCE & OPTIMIZATION
# ============================================================================
cython==3.0.6
numba==0.58.1

# ============================================================================
# ADDITIONAL UTILITIES
# ============================================================================
tqdm==4.66.1
click==8.1.7
rich==13.7.0

# ============================================================================
# REFERENCE ARCHITECTURE INTEGRATIONS (Added Jan 2026)
# ============================================================================
# Multi-exchange crypto support (107+ exchanges)
ccxt>=4.2.0

# Standard RL environment + research-grade algorithms
gymnasium>=1.0.0
stable-baselines3>=2.3.0

# Unified financial data layer (350+ datasets)
openbb>=4.3.0

# LangGraph agent orchestration
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
